Separate lab config from server launch config
